Implemented Order property to cart products

+++ Shared/Models/CartProduct/CartProduct.cs
@@ -5,5 +5,5 @@ namespace GroceryListHelper.Shared.Models.CartProduct;
 public class CartProduct : StoreProductModel
 {
     public double Amount { get; set; }
-    public int Order { get; set; }
+    public double Order { get; set; }
 }
